seperate the strategy in to individual files
